Added characteristic validation
